Move changes-sync logic out of josh-cli into shared crates
Factor `josh changes sync` into a reusable in-process entry point so
both the CLI and other consumers (the GUI) can call it without shelling
out. No CLI behavior change.

- `josh-changes` gains `remote_config` (Forge, RemoteConfig,
  read/write/migrate helpers moved out of `josh-cli`) and a small
  `sync_local` wrapper around the existing `sync_changes`.
- `josh-github-changes` gains `connection` (make_api_connection,
  api_connection_hint moved from `josh-cli/src/forge/github.rs`) and
  `sync` (`SyncOptions`, `SyncReport`, async `sync_from_github`, plus
  a dispatcher `sync` that handles both Local and Remote scopes).
- `josh-cli/src/commands/sync.rs` collapses to ScopeArgs resolution +
  `block_on(josh_github_changes::sync::sync(...))`. The old in-function
  `tokio::runtime::Runtime::new()` is gone; the caller drives it now.
- The old `config.rs` and `forge/{mod,github}.rs` in `josh-cli` become
  thin re-export shims so other CLI call sites keep compiling without
  per-site changes.

Change: factor-changes-sync-into-josh-github-changes
