+++ forges/josh-github-changes/src/lib.rs
@@ -1,8 +1,10 @@
 pub mod admission;
 mod comments;
+pub mod connection;
 mod display;
 mod prs;
 pub mod repo;
+pub mod sync;
 
 pub use comments::{
     post_local_comments, post_local_votes, sync_change_comments, sync_change_comments_by_pr_number,
