+++ josh-changes/Cargo.toml
@@ -11,7 +11,9 @@ keywords = ["git", "monorepo", "workflow", "scm"]
 [dependencies]
 anyhow.workspace = true
 chrono.workspace = true
+clap.workspace = true
 git2.workspace = true
+log.workspace = true
 serde.workspace = true
 serde_json.workspace = true
